% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devMSMsHelpers-package.R
\name{devMSMHelpers_common_docs}
\alias{devMSMHelpers_common_docs}
\title{Common documentation for devMSMs helper functions}
\usage{
devMSMHelpers_common_docs(
  data,
  obj,
  sep,
  exposure,
  outcome,
  factor_confounders,
  integer_confounders,
  hi_lo_cut,
  reference,
  comparison,
  verbose,
  save.out,
  home_dir
)
}
\arguments{
\item{data}{data in long format or in wide format as: a data frame, list of imputed data
frames, or \code{mids} object from the \code{mice} package}

\item{obj}{initialized MSM object from \code{\link[=initMSM]{initMSM()}}}

\item{sep}{(optional) delimiter for time integer; must be in regex format,
e.g., "\\." (default is period)}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{factor_confounders}{(optional) list of variable names that should be made into factors
(default is numeric)}

\item{integer_confounders}{(optional) list of variable names that should be made into integers
(default is numeric)

@keywords internal}

\item{hi_lo_cut}{list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure
(default is median split)}

\item{reference}{lists of one or more strings of "-"-separated "l"
and "h" values indicative of a reference exposure history to which to
compare comparison, required if comparison is supplied}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is supplied}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}

\item{home_dir}{path to home directory (required if save.out = TRUE)}
}
\description{
Common documentation for devMSMs helper functions
}
