% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devMSMsHelpers-package.R
\docType{package}
\name{devMSMsHelpers-package}
\alias{devMSMsHelpers}
\alias{devMSMsHelpers-package}
\title{devMSMsHelpers: Helper functions to accompany the devMSMs package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Isabella Stallworthy \email{istall@seas.upenn.edu}

Authors:
\itemize{
  \item Meriah DeJoseph \email{meriahd@stanford.edu}
  \item Emily Padrutt \email{padru004@umn.edu}
  \item Noah Greifer \email{ngreifer@iq.harvard.edu}
  \item Daniel Berry \email{dberry@umn.edu}
}

}
\keyword{internal}
