% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_eval_hist.R
\name{print_eval_hist}
\alias{print_eval_hist}
\title{Visualize distribution of sample across exposure histories}
\usage{
print_eval_hist(
  epoch_history,
  epoch,
  hi_lo_cut,
  reference = NULL,
  comparison = NULL
)
}
\arguments{
\item{epoch_history}{character vector of length \code{nrow(data)} containing epochi histories for
each unit}

\item{epoch}{character vector of epoch values}
}
\value{
invisibly NULL
}
\description{
Create customized, user-specified exposure histories and tables displaying
sample distribution across them for user inspection.
}
\keyword{internal}
