% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_var}
\alias{summarize_var}
\title{Summarize variable}
\usage{
summarize_var(
  var,
  data,
  format,
  sep,
  exposure_time_pts = NULL,
  print = TRUE,
  home_dir = NULL,
  save = TRUE
)
}
\description{
Summarize variable
}
\keyword{internal}
