% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectData.R
\name{inspectData}
\alias{inspectData}
\title{Inspect long/wide/imputed data}
\usage{
inspectData(
  data,
  obj,
  outcome,
  sep = "\\\\.",
  hi_lo_cut = NULL,
  reference = NULL,
  comparison = NULL,
  verbose = TRUE,
  save.out = TRUE
)
}
\arguments{
\item{data}{data in wide format as: a data frame, list of imputed data
frames, or mids object}

\item{outcome}{name of outcome variable with ".timepoint" suffix}

\item{hi_lo_cut}{(optional) list of two numbers indicating quantile values
that reflect high and low values, respectively, for continuous exposure
(default is median split)}

\item{reference}{(optional) list of one or more strings of "-"-separated "l" and "h" values
indicative of a reference exposure history to which to compare comparison,
required if comparison is specified}

\item{comparison}{(optional) list of one or more strings of "-"-separated "l"
and "h" values indicative of comparison history/histories to compare to
reference, required if reference is specified}

\item{verbose}{(optional) TRUE or FALSE indicator for user output (default is
TRUE)}

\item{save.out}{(optional) TRUE or FALSE indicator to save output and
intermediary output locally (default is TRUE)}

\item{home_dir}{(optional) path to home directory (required if save.out = TRUE)}
}
\value{
none
}
\description{
Inspect long/wide/imputed data
}
\examples{
test <- data.frame(ID = 1:50,
                   A.1 = rnorm(n = 50),
                   A.2 = rnorm(n = 50),
                   A.3 = rnorm(n = 50),
                   B.1 = rnorm(n = 50),
                   B.2 = rnorm(n = 50),
                   B.3 = rnorm(n = 50),
                   C = rnorm(n = 50),
                   D.3 = rnorm(n = 50))
test[, c("A.1", "A.2", "A.3")] <- lapply(test[, c("A.1", "A.2", "A.3")], as.numeric)

inspectData(data = test,
            exposure = c("A.1", "A.2", "A.3"),
            outcome = "D.3",
            save.out = FALSE)
inspectData(data = test,
            exposure = c("A.1", "A.2", "A.3"),
            outcome = "D.3",
            hi_lo_cut = c(0.8, 0.2),
            save.out = FALSE)
inspectData(data = test,
            exposure = c("A.1", "A.2", "A.3"),
            outcome = "D.3",
            hi_lo_cut = c(0.8, 0.2),
            reference = "l-l-l",
            comparison = "h-h-h",
            save.out = FALSE)
}
